WebCL-conformance/conformance/bindingTesting/contextAndCreate/cl_buffer_createSubBuffer.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLBuffer createSubBuffer conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCLBuffer :: createSubBuffer test case.");
    debug("(CLenum memFlags, CLuint origin, CLuint sizeInBytes)");
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, BUFFER_SIZE);
    var origin = 0, subBufferSize = 10;

    shouldBeType("webCLBuffer.createSubBuffer(webcl.MEM_READ_ONLY, origin, subBufferSize);", WebCLBuffer);
    shouldBeType("webCLBuffer.createSubBuffer(webcl.MEM_WRITE_ONLY, origin, subBufferSize);", WebCLBuffer);
    shouldBeType("webCLBuffer.createSubBuffer(webcl.MEM_READ_WRITE, origin, subBufferSize);", WebCLBuffer);

    // Testing for exceptions.
    var originInvalid = BUFFER_SIZE + 1;
    shouldThrowExceptionName("webCLBuffer.createSubBuffer(webcl.MEM_READ_ONLY, originInvalid, subBufferSize);", "INVALID_VALUE");

    var sizeInBytesInvalid = BUFFER_SIZE + 1;
    shouldThrowExceptionName("webCLBuffer.createSubBuffer(webcl.MEM_READ_ONLY, origin, sizeInBytesInvalid);", "INVALID_VALUE");

    shouldThrowExceptionName("webCLBuffer.createSubBuffer(invalid_CLenum, origin, subBufferSize);", "INVALID_VALUE");

    var webCLBufferRead = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    shouldThrowExceptionName("webCLBufferRead.createSubBuffer(webcl.MEM_WRITE_ONLY, origin, subBufferSize);", "INVALID_VALUE");
    shouldThrowExceptionName("webCLBufferRead.createSubBuffer(webcl.MEM_READ_WRITE, origin, subBufferSize);", "INVALID_VALUE");

    var webCLBufferWrite = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    shouldThrowExceptionName("webCLBufferWrite.createSubBuffer(webcl.MEM_READ_ONLY, origin, subBufferSize);", "INVALID_VALUE");
    shouldThrowExceptionName("webCLBufferWrite.createSubBuffer(webcl.MEM_READ_WRITE, origin, subBufferSize);", "INVALID_VALUE");

    var webCLSubBuffer = wtu.createSubBuffer(webCLBuffer, webcl.MEM_READ_WRITE, origin, 100);
    shouldThrowExceptionName("webCLSubBuffer.createSubBuffer(webcl.MEM_READ_WRITE, origin, subBufferSize);", "INVALID_MEM_OBJECT");

    var contextDeviceArray = webCLContext.getInfo(webcl.CONTEXT_DEVICES);
    var unalignedOrigin = null;
    if (contextDeviceArray.length > 1) {
        for (i = 0; i < contextDeviceArray.length - 1; i++) {
            if (contextDeviceArray[i].getInfo(webcl.DEVICE_MEM_BASE_ADDR_ALIGN) !=
                contextDeviceArray[i+1].getInfo(webcl.DEVICE_MEM_BASE_ADDR_ALIGN)) {
                unalignedOrigin = contextDeviceArray[i+1].getInfo(webcl.DEVICE_MEM_BASE_ADDR_ALIGN);
                break;
            }
        }
        if (unalignedOrigin)
            shouldThrowExceptionName("webCLBuffer.createSubBuffer(webcl.MEM_READ_ONLY, unalignedOrigin, subBufferSize);", "MISALIGNED_SUB_BUFFER_OFFSET");
    }

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
